% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/bparrows.R
\name{bparrows}
\alias{bparrows}
\title{Biplot Arrows Function}
\usage{
bparrows(x, myscale = 1, matrix.name, ...)
}
\arguments{
\item{x}{pca species scores object.}

\item{myscale}{scale of arrows to be plotted Defaults to 1.}

\item{matrix.name}{name of data matrix.}
}
\description{
Biplot Arrows Function
}
\examples{
bparrows()
}
\keyword{vegan}

