% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/bub.R
\name{bub}
\alias{bub}
\title{Bubble Function}
\usage{
bub(x, expand = 5)
}
\arguments{
\item{x}{datamatrix.}

\item{expand}{size of bubbles Defaults to 5.}
}
\description{
Bubble Function
}
\examples{
bub()
}
\keyword{vegan}

