\name{ffg.cast2}
\alias{ffg.cast2}
\docType{data}
\title{Raw data version of ffg.cast

}
\description{
Raw macroinvertebrate functional feeding group biomass data from Slab
Camp Creek Kentucky.
}
\usage{data("ffg.cast2")}
\format{
  A data frame with 18 observations on the following 10 variables.
  \describe{
    \item{\code{SAMPID}}{a character vector}
    \item{\code{STREAM}}{a factor with levels \code{SC}}
    \item{\code{SEASON}}{a character vector}
    \item{\code{CU}}{a factor with levels \code{pool} \code{riffle}}
    \item{\code{YR.POSTREST}}{a factor with levels \code{1} \code{2} \code{3}}
    \item{\code{COLLECTOR.FILTERER}}{a numeric vector}
    \item{\code{COLLECTOR.GATHERER}}{a numeric vector}
    \item{\code{HERBIVORE}}{a numeric vector}
    \item{\code{PREDATOR}}{a numeric vector}
    \item{\code{SHREDDER}}{a numeric vector}
  }
}
\examples{
data(ffg.cast2)
}
\keyword{datasets}
