% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/levels.to.num.R
\name{levels.to.num}
\alias{levels.to.num}
\title{Convert levels of type character into numeric}
\usage{
\method{levels}{to.num}(x)
}
\arguments{
\item{x}{Vector to perform the function.}
}
\description{
Convert levels of type character into numeric
}
\examples{
data(ffg.cast)
levels.to.num(ffg.cast$CU)
}
\keyword{vegan}

