% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/nmds.plot.R
\name{nmds.plot}
\alias{nmds.plot}
\title{Custom NMDS Plotting Function}
\usage{
nmds.plot(nmds.object, col.factor, pch.factor, main = NULL, myscale = 1,
  cex = 1)
}
\arguments{
\item{nmds.object}{NMDS object from metaMDS function (vegan).}

\item{col.factor}{Name of vector to be used to color the points.}

\item{pch.factor}{Name of vector to be used to create symbols.}

\item{main}{Title of plot.}

\item{myscale}{Scaling size for bparrows Defaults to 0.25.}

\item{cex}{Point size Defaults to 1.}
}
\description{
Custom NMDS Plotting Function
}
\examples{
data(ffg.cast)
ffg.bray <- vegdist(ffg.cast[,-1:-4])
nmds.ffg <- metaMDS(ffg.bray)
nmds.plot(nmds.ffg, ffg.cast$CU, ffg.cast$CU, main='NMDS Plot 1')
}
\keyword{vegan}

