% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/pca.reg.R
\name{pca.reg}
\alias{pca.reg}
\title{Linear Regression on PCA objects}
\usage{
pca.reg(pca.object, group, matrix.name, myscale, plot = TRUE)
}
\arguments{
\item{pca.object}{PCA object from rda function (vegan).}

\item{group}{Name of vector to be used to specify groups within the data.}

\item{matrix.name}{Name of the matrix containing the data.}

\item{myscale}{Scaling size for bparrows Defaults to 0.25.}

\item{plot}{Plot the regression lines Defaults to TRUE.}
}
\description{
This functions allows one to perform a linear regression on groups of PCA points to determine the trajectory of points within a specified group.
The regression lines for each group are then used to determine the correlation values between these lines and  all species vectors which are beyond the circle of equlibrium.
The output is a dataframe containing correlation values for each group when compared to each species vector.
}
\examples{
data(ffg.cast)
ffg.hel <- decostand(ffg.cast[,-1:-4], 'hellinger')
pca.hel <- rda(ffg.hel)
pca.hel.scores <- scores(pca.hel, display='sites', scaling=1)
pca.hel.plot <- pca.plot(pca.hel, ffg.cast$CU, ffg.cast$CU, ffg.hel, cex=1.5)
text(pca.hel.scores, labels=ffg.cast$YR.POSTREST, pos=3)
lines(pca.hel.scores[-4:-6,], col='red')
lines(pca.hel.scores[-1:-3,], col='green')
pca.reg(pca.hel, ffg.cast$CU, ffg.cast, 0.25, plot=TRUE)


}
\keyword{vegan}

