% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/pca.reg2.R
\name{pca.reg2}
\alias{pca.reg2}
\title{Linear Regression on PCA objects 2}
\usage{
pca.reg2(pca.object, sp.name, group, plot = TRUE)
}
\arguments{
\item{pca.object}{PCA object from rda function (vegan).}

\item{sp.name}{Name of species vector of interest.}

\item{group}{Name of vector to be used to specify groups within the data.}

\item{plot}{Plot the regression lines Defaults to TRUE.}
}
\description{
This functions allows one to perform a linear regression on groups of
PCA points to determine the trajectory or trend of points within a specified group.
The regression lines for each group are then used to determine the correlation values between these lines and a specified species vector.
The output is a dataframe containing correlation values for each group when compared to the specified species vector.
}
\examples{
data(ffg.cast)
ffg.hel <- decostand(ffg.cast[,-1:-4], 'hellinger')
pca.hel <- rda(ffg.hel)
pca.hel.scores <- scores(pca.hel, display='sites', scaling=1)
pca.hel.plot <- pca.plot(pca.hel, ffg.cast$CU, ffg.cast$CU, ffg.hel, cex=1.5)
text(pca.hel.scores, labels=ffg.cast$YR.POSTREST, pos=3)
lines(pca.hel.scores[-4:-6,], col='red')
lines(pca.hel.scores[-1:-3,], col='green')
pca.reg2(pca.hel, 'SHREDDER', ffg.cast$CU, plot=TRUE)
}
\keyword{vegan}

