% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/pro.eig.R
\name{pro.eig}
\alias{pro.eig}
\title{A Function to Allow the Computation of the Proportion of Variance Explained By Each Eigenvalue.}
\usage{
pro.eig(pca.object)
}
\arguments{
\item{pca.object}{an object resulting from the rda function (vegan).}
}
\description{
A Function to Allow the Computation of the Proportion of Variance Explained By Each Eigenvalue.
}
\examples{
pro.eig()
}
\keyword{vegan}

